Website/requirements.txt
Flask==1.1.1
Flask-Mail==0.9.1
Flask-WTF==0.14.3
WTForms==2.3.1
PyMySQL
email-validator==1.1.1
gunicorn
pyjwt
python-dotenv